# Block RNGs: remove unaligned memory cast