Fix #779. Review please @RalfJung.

The point of the removed specialisations was to avoid one copy. Since the output type may not have the correct alignment and we wish to copy bytes in the same order, we have no choice but to use a buffer anyway. (We could complicate things by checking the alignment at run-time, but I don't think it's worthwhile given the very small performance cost of the extra copy.)

Benchmarking with 10*1024 byte buffer does show a small impact:
```
# before
test gen_bytes_chacha20             ... bench:  17,775,692 ns/iter (+/- 290,563) = 576 MB/s
test gen_bytes_hc128                ... bench:   4,243,890 ns/iter (+/- 69,032) = 2412 MB/s
test gen_bytes_isaac                ... bench:   7,238,757 ns/iter (+/- 123,797) = 1414 MB/s
test gen_bytes_isaac64              ... bench:   3,900,653 ns/iter (+/- 47,162) = 2625 MB/s
# after
test gen_bytes_chacha20             ... bench:  18,007,036 ns/iter (+/- 384,358) = 568 MB/s
test gen_bytes_hc128                ... bench:   4,634,505 ns/iter (+/- 55,119) = 2209 MB/s
test gen_bytes_isaac                ... bench:   7,338,592 ns/iter (+/- 90,396) = 1395 MB/s
test gen_bytes_isaac64              ... bench:   4,066,367 ns/iter (+/- 55,543) = 2518 MB/s
```

The second commit cleans up some warnings in the benches.