@burdges 2019-04-22
It's not worth it to use `read_unaligned` like you mentioned previously?