@dhardy 2019-04-22
This isn't about unaligned reads, it's about unaligned writes masquerading as aligned ones (because of cast to `&[u32]` or `&[u64]`). I guess we could change `BlockRngCore::generate` to take `*mut u32` arg or some such, but I really don't think these small performance losses are worth pushing `unsafe` code into all block-rng implementations.