@RalfJung 2019-04-22
Is there a way that [`align_to`](https://doc.rust-lang.org/nightly/std/primitive.slice.html#method.align_to) could be useful here to find the "aligned" part of the buffer?  That would work on all platforms then, not just x86.

The change itself seems fine; this is what Miri is testing in https://github.com/rust-random/rand/pull/781.