@dhardy 2019-04-22
Finding the aligned subset of `dest` doesn't help unless we allow bytes of the RNG to be skipped. We don't wish to do this arbitrarily (or based on something as tenuous as input alignment) because in some cases reproducibility is important, and this would be hard to document and a breaking change from past behaviour.

I'll leave this until tomorrow for any further review, then we can merge the PRs.