@vks 2019-04-23
> Finding the aligned subset of dest doesn't help unless we allow bytes of the RNG to be skipped.

Maybe we should add an API that requires aligned slices?