@dhardy 2019-04-23
As in `fill_aligned_bytes`? Possible I guess.

What I don't understand is why everyone keeps suggesting complicated ways to keep this (very) small highly-specific optimisation.