@burdges 2019-04-23
It's not worth complicating the trait over this.  lol