@LukasKalbertodt 2019-07-27
Would it be possible to backport this bugfix to `rand_core 0.4`? Currently it's not possible to run [quickcheck](https://crates.io/crates/quickcheck) tests through Miri because this bug always makes the tests fail. And I think quite a few crates have not switched to the new rand version yet. 

But I can also understand if this is too much work. I just wanted to ask ^_^