@dhardy 2019-07-27
Sure, I guess it's possible.